requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
brotlipy==0.7.0
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
idna==2.10
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
pycares==4.0.0
pycodestyle==2.7.0
pycparser==2.20
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3